requirements.txt
os
discord.py
pickle
openai
python-dotenv
